Backend/requirements.txt
google-generativeai>=0.3.0
PyPDF2>=3.0.0
python-dotenv>=1.0.0
fpdf2>=2.7.5
